Gave in and exposed state to aid in testing
Originally the code was written to keep the data as close as possible
to the execution to try and eke out a bit more performance.  Though
there were likely some small perf advantages, unit testing would have
been next to impossible.

The danger is that now the internal state is now exposed as well.
State is not an interface!  It can/will change.