Simplified and fixed buffer logic
Completely reworked the buffer algorithm with a much simpler and less
error prone approach.  Items are not reliably moved one at a time to
the front/end of the list where scrolling back was previously leaving
holes in the view.

Wrote a visualization page to make it easier to see what is happening
as the user scrolls up/down on the page.